Share the backend surface the three terminals implemented separately
Output accumulation, the reply-failure reporting, the cancellation check and
the owner constants were byte-identical in each backend; they now live on
TerminalProcess and in terminal_process/terminal_queries, and the backends keep
only their read loops. The terminal child environment and the reply-resolution
mapping become one shared function each.
